requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
cssbeautifier==1.14.8
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
djlint==1.31.0
docopt==0.6.2
dominate==2.8.0
EditorConfig==0.12.3
executing==1.2.0
filelock==3.12.2
Flask
flask-appconfig==0.11.1
flask-nav==0.6
frozenlist==1.3.3
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
ipykernel==6.23.2
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsbeautifier==1.14.8
json5==0.9.14
jupyter_client==8.2.0
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
openai==0.27.8
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
pip==23.1.2
pipenv==2023.6.12
pipreqs==0.4.13
platformdirs==3.5.3
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
setuptools==67.8.0
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
urllib3==2.0.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
visitor==0.1.3
wcwidth==0.2.6
Werkzeug==2.3.6
wheel==0.40.0
yarg==0.1.9
yarl==1.9.2
